# src/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: Inventory Management API
  version: 1.0.0
  description: API documentation for Inventory Management system.
servers:
  - url: http://localhost:3001
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []
...
paths:
...
  /products:
    get:
      tags: [Products]
      summary: Get products list
      parameters:
        - name: search
          in: query
          schema: { type: string }
      responses:
        '200':
          description: List of products
    post:
      tags: [Products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, price]
              properties:
                name: { type: string }
                price: { type: number }
                stockQuantity: { type: number }
                rating: { type: number }
      responses:
        '201': { description: Created }
...
  /users:
    get:
      tags: [Users]
      summary: Get users list (Admin only)
      responses:
        '200': { description: List of users }
    post:
      tags: [Users]
      summary: Create user (Admin only)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                email: { type: string }
                role: { type: string }
      responses:
        '201': { description: Created }
...
  /expenses:
    post:
      tags: [Expenses]
      summary: Create expense
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                category: { type: string }
                amount: { type: number }
                date: { type: string }
      responses:
        '201': { description: Created }
...
  /invoices:
    get:
      tags: [Invoices]
      summary: Get invoices
      responses:
        '200': { description: List of invoices }
    post:
      tags: [Invoices]
      summary: Create invoice
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                customerName: { type: string }
                amount: { type: number }
                dueDate: { type: string }
                items: 
                  type: array
                  items:
                    type: object
                    properties:
                      productId: { type: string }
                      quantity: { type: number }
                      unitPrice: { type: number }
      responses:
        '201': { description: Created }
...
  /purchases:
    get:
      tags: [Purchases]
      summary: Get purchases
      responses:
        '200': { description: List of purchases }
    post:
      tags: [Purchases]
      summary: Create purchase
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                supplierId: { type: string }
                products:
                  type: array
                  items:
                    type: object
                    properties:
                      productId: { type: string }
                      quantity: { type: number }
                      unitCost: { type: number }
      responses:
        '201': { description: Created }
...